+++ index.d.ts
@@ -228,7 +228,6 @@ declare module 'pc-nrfconnect-shared' {
     }
 
     interface Device extends Omit<NrfdlDevice, 'usb' | 'jlink'> {
-        // serialnumber?: string; // from nrf-device-lib
         // traits: DeviceTraits; // from nrf-device-lib
         // serialports?: Serialport[]; // from nrf-device-lib
         usb?: {
@@ -237,6 +236,12 @@ declare module 'pc-nrfconnect-shared' {
         jlink?: {
             boardVersion: string;
         }; // from nrf-device-lib
+        /**
+         * @deprecated Using the property `comName` has been
+         * deprecated. You should now use `path`. The property
+         * will be removed in the next major release.
+         */
+        serialnumber?: string; // from nrf-device-lib
         serialNumber: string;
         boardVersion?: string;
         nickname?: string;