+++ Changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Exposes deviceInfo
 ### Added
 - Component `StateSelector`.
+- Added `Dropdown` component.
 
 ## 4.28.2 - 2021-07-19
 ### Changed